Fix deploys
Move the deploy steps to their own jobs that only get run on main and
production. The previous setup wouldn't deploy, but you would see that
step in the job "pass". The job was "successful" in skipping over the
step when being run on a non-deployable branch.

Kinda startling to read a PR and realize that Actions automatically ran
your deploy step for a stranger. This fixes all that.

Remove some trailing whitespace

Move the deploy step to it's own job

who ammm i when deploying

De-specify the workflow file name

Use the deploy action

we got uses instead

Split and name steps

bring back runs on

Deploy to production too